feat(populatedb): populate database with script (#12)
* feat: display all messages on the home route

* refactor(pool): do not use connection string to connect to Neon-hosted PostgreSQL database - instead, follow how it is done in the documentation

see link: https://neon.com/docs/guides/node#configure-the-postgres-client

* feat(populatedb): populate local database with script (#11)

* feat(populatedb): add SQL query for adding four columns, and then populate with two rows of test data

for 'added_date_time' i've followed postgres' documentation here: https://www.postgresql.org/docs/current/sql-altertable.html#id-1.9.3.35.8 . this is because i do want the time zone and i want to have it default to the 'now' when the row is added. the two rows are added as test data

* refactor(pool): use connection URI to defining connection information

this is so that I can actually use the same environment variable in my populatedb script

* feat(populatedb): add script to create table and add some pre-populated values to it

* feat(package.json): add npm script to seed database more easily

* feat: when seeding, URL of database must be provided, otherwise error is thrown

this allows user to seed either local db or prod db as required. however, by default, app connects to prod db in the pg.pool